Add edit and detail view

+++ HomeTask/Data/DataStore/StockDataStore.swift
@@ -13,7 +13,7 @@ protocol StockDataStore {
     func deleteStock(id: UUID)
     func incrementAmount(id: UUID)
     func decrementAmount(id: UUID)
-    func update(id: UUID, title: String, memo: String?, amount: Int, limit: Int)
+    func update(id: UUID, title: String, memo: String?, amount: Int, limit: Double)
 }
 
 struct StockDataStoreImpl: StockDataStore {
@@ -37,7 +37,7 @@ struct StockDataStoreImpl: StockDataStore {
         StockEntity.decrement(id: id)
     }
     
-    func update(id: UUID, title: String, memo: String?, amount: Int, limit: Int) {
+    func update(id: UUID, title: String, memo: String?, amount: Int, limit: Double) {
         StockEntity.update(id: id, title: title, memo: memo, amount: amount, limit: limit)
     }
 }
