+++ HomeTask/Data/Repository/StockRepositoy.swift
@@ -13,7 +13,7 @@ protocol StockRepository: AnyObject {
     func delete(id: UUID)
     func increment(id: UUID)
     func decrement(id: UUID)
-    func update(id: UUID, title: String, memo: String?, amount: Int, limit: Int)
+    func update(id: UUID, title: String, memo: String?, amount: Int, limit: Double)
 }
 
 class StockRepositoryImpl: StockRepository {
@@ -40,7 +40,7 @@ class StockRepositoryImpl: StockRepository {
         dataStore.decrementAmount(id: id)
     }
     
-    func update(id: UUID, title: String, memo: String?, amount: Int, limit: Int) {
+    func update(id: UUID, title: String, memo: String?, amount: Int, limit: Double) {
         dataStore.update(id: id, title: title, memo: memo, amount: amount, limit: limit)
     }
 }
